feat: api get bot source extracted data
