+++ src/server/api/root.ts
@@ -1,8 +1,9 @@
 import { postRouter } from '~/server/api/routers/post'
 import { createCallerFactory, createTRPCRouter } from '~/server/api/trpc'
 import { botRouter } from './routers/bot'
-import { botIntegrationRouter } from './routers/bot_integration'
-import { botSourceRouter } from './routers/bot_source'
+import { botIntegrationRouter } from './routers/bot-integration'
+import { botSourceRouter } from './routers/bot-source'
+import { botSourceExtractedDataRouter } from './routers/bot-source-extracted-data'
 import { chatRouter } from './routers/chat'
 
 /**
@@ -14,6 +15,7 @@ export const appRouter = createTRPCRouter({
   bot: botRouter,
   botIntegrationRouter: botIntegrationRouter,
   botSource: botSourceRouter,
+  botSourceExtractedDataRouter: botSourceExtractedDataRouter,
   chatRouter: chatRouter,
   post: postRouter,
 })
